job-agent-gemini/Langchain_Local/requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai